Extract SIMD math into bitnet-simd microcrate
