+++ crates/bitnet-simd/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "bitnet-simd"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "SIMD-accelerated vector math primitives for BitNet"
+rust-version.workspace = true
+
+[dependencies]
+
+[features]
+default = []
+
+[lints]
+workspace = true
